nav: Move from lit -> react
