+++ packages/data/custom.d.ts
@@ -9,12 +9,6 @@ declare module "*.html" {
   export default content;
 }
 
-declare namespace JSX {
-  interface IntrinsicElements {
-    "cv-nav": any;
-  }
-}
-
 declare module "!!raw-loader!*" {
   const contents: string;
   export = contents;
