+++ packages/nav/package.json
@@ -16,7 +16,7 @@
   },
   "devDependencies": {
     "@types/node": "^18.11.9",
-    "es-dev-server": "^2.1.0",
+    "@types/react": "^18.0.25",
     "raw-loader": "^4.0.2",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.3"
@@ -28,7 +28,6 @@
     "eslint-plugin-import": "^2.25.1"
   },
   "dependencies": {
-    "@webcomponents/webcomponentsjs": "^2.7.0",
-    "lit": "^2.4.1"
+    "react": "^18.2.0"
   }
 }
