+++ packages/react/src/custom.d.ts
@@ -16,7 +16,6 @@ declare module "*.html" {
 
 declare namespace JSX {
   interface IntrinsicElements {
-    "cv-nav": any;
     "range-slider": any;
     "simple-switcher": any;
   }
